+++ vpc/include/tencentcloud/vpc/v20170312/model/BandwidthPackageBillBandwidth.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2017-2019 THL A29 Limited, a Tencent company. All Rights Reserved.
+ * Copyright (c) 2017-2025 Tencent. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -49,18 +49,21 @@ namespace TencentCloud
                     /**
                      * 获取当前计费用量，单位为 Mbps
                      * @return BandwidthUsage 当前计费用量，单位为 Mbps
+                     * 
                      */
-                    uint64_t GetBandwidthUsage() const;
+                    double GetBandwidthUsage() const;
 
                     /**
                      * 设置当前计费用量，单位为 Mbps
-                     * @param BandwidthUsage 当前计费用量，单位为 Mbps
+                     * @param _bandwidthUsage 当前计费用量，单位为 Mbps
+                     * 
                      */
-                    void SetBandwidthUsage(const uint64_t& _bandwidthUsage);
+                    void SetBandwidthUsage(const double& _bandwidthUsage);
 
                     /**
                      * 判断参数 BandwidthUsage 是否已赋值
                      * @return BandwidthUsage 是否已赋值
+                     * 
                      */
                     bool BandwidthUsageHasBeenSet() const;
 
@@ -69,7 +72,7 @@ namespace TencentCloud
                     /**
                      * 当前计费用量，单位为 Mbps
                      */
-                    uint64_t m_bandwidthUsage;
+                    double m_bandwidthUsage;
                     bool m_bandwidthUsageHasBeenSet;
 
                 };
